- Extracted out the API request and response models out of the HSCatalog class. - Reintroduced the Type xp for catalogXp. Renamed to UserGroupCatalogXp for better context.
