+++ src/Middleware/src/Headstart.Common/Models/Headstart/HSCatalog.cs
@@ -1,26 +1,16 @@
-﻿using System.Collections.Generic;
-using OrderCloud.SDK;
+﻿using OrderCloud.SDK;
 
 namespace Headstart.Models
 {
-    public class HSCatalog : UserGroup<CatalogXp>, IHSObject
+    public class HSCatalog : UserGroup<UserGroupCatalogXp>, IHSObject
     {
     }
 
-    // potentially use this for the api later
-    public class HSCatalogAssignment
+    public class UserGroupCatalogXp
     {
-        public string LocationID { get; set; }
-
-        public string CatalogID { get; set; }
-    }
-
-    public class HSCatalogAssignmentRequest
-    {
-        public List<string> CatalogIDs { get; set; }
-    }
-
-    public class CatalogXp
-    {
+        /// <summary>
+        /// Provides context type to the user group to drive functionality. User groups created with this xp will always be of type "Catalog".
+        /// </summary>
+        public string Type { get; set; } = "Catalog";
     }
 }
