+++ src/Middleware/src/Headstart.Common/Models/Responses/HSCatalogAssignmentResponse.cs
@@ -1,0 +1,9 @@
+﻿namespace Headstart.Common.Models
+{
+    public class HSCatalogAssignmentResponse
+    {
+        public string LocationID { get; set; }
+
+        public string CatalogID { get; set; }
+    }
+}
